+++ packages/stream_chat_flutter/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 - Updated Dart SDK constraints to `>=2.14.0 <3.0.0`
 
+✅ Added
+
+- `MessageListViewThemeData` now accepts a `DecorationImage` as a background image for `MessageListView`.
+
 ## 3.1.1
 
 - Updated `stream_chat_flutter_core` dependency to [`3.1.1`](https://pub.dev/packages/stream_chat_flutter_core/changelog).